Creates new app for Comment model
