+++ insta_comment/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from insta_comment.models import Comment
+
+admin.site.register(Comment)
