+++ insta_post/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from insta_post.models import FavoriteCar, Comment
+from insta_post.models import FavoriteCar
 
 admin.site.register(FavoriteCar)
-admin.site.register(Comment)
+
