+++ instacar/urls.py
@@ -3,8 +3,9 @@ from django.urls import path
 from django.conf import settings
 from django.conf.urls.static import static
 from authentication.views import signup_view, logout_view, LoginView
-from insta_user.views import index, profile_view, profile_edit_view, del_user
+from insta_user.views import profile_edit_view, del_user, profile_view
 from insta_post import views
+from insta_comment.views import comment_form_view, del_comment, edit_comment, comment_likes
 from django.views.generic import TemplateView
 from django.views.defaults import page_not_found, server_error
 
@@ -17,12 +18,12 @@ urlpatterns = [
     path('newpost/', views.post_form_view),
     path('upvote/<int:post_id>/', views.up_vote),
     path('post/<int:post_id>/', views.photo_detail, name="post"),
-    path('post/<int:post_id>/newcomment/', views.comment_form_view),
+    path('post/<int:post_id>/newcomment/', comment_form_view),
     path('post/<int:post_id>/edit/', views.post_edit_view, name='post_edit'),
     path('post/<int:post_id>/delete/', views.del_post, name='del_post'),
-    path('comment/<int:pk>/delete/', views.del_comment, name='del_comment'),
-    path('comment/<int:pk>/edit/', views.edit_comment, name='del_comment'),
-    path('comment/<int:pk>/like/', views.comment_likes, name='like'),
+    path('comment/<int:pk>/delete/', del_comment, name='del_comment'),
+    path('comment/<int:pk>/edit/', edit_comment, name='del_comment'),
+    path('comment/<int:pk>/like/', comment_likes, name='like'),
     path('following/<int:follow_id>/', views.FollowView.as_view()),
     path('unfollowing/<int:unfollow_id>/', views.UnfollowView.as_view()),
     path('500/', TemplateView.as_view(template_name="500.html")),
